+++ backend/app/app/models/preburn_tx.py
@@ -1,5 +1,5 @@
 import enum
-from sqlalchemy import Column, Integer, String, Enum
+from sqlalchemy import Column, Integer, String, Enum, Float
 
 from app.db.base_class import Base
 
@@ -13,7 +13,7 @@ class PBTXStatus(enum.Enum):
 
 class PreBurnTx(Base):
     id = Column(Integer, primary_key=True, index=True)
-    fee = Column(Integer, index=True)
+    fee = Column(Float, index=True)
     height = Column(Integer, index=True)
     txid = Column(String)
     status = Column(Enum(PBTXStatus), default=PBTXStatus.NEW)
