+++ backend/app/app/schemas/preburn_tx.py
@@ -6,7 +6,7 @@ from app.models.preburn_tx import PBTXStatus
 
 # Shared properties
 class PreBurnTxBase(BaseModel):
-    fee: int
+    fee: float
     height: int
     txid: str
